✨ feat(search_form.dart): add NikSearchForm widget This commit adds a new widget called NikSearchForm which is used to search for a National Identification Number (NIK). The widget contains a text field that accepts numeric input only and has a custom Google font applied to it. The text field also has validation to ensure that it is not empty.
✨ feat(search_result.dart): add NikSearchResult widget
This commit adds the NikSearchResult widget to the project. The widget is used to display search results for a particular NIK. It contains a ListView builder that displays a list of debitur search results. The widget also contains a Card widget that displays the details of each debitur. The details include the debitur's name, NIK, jenis usaha, and tgl input. The widget also contains a RandomAvatar widget that displays a random avatar for each debitur. Finally, the widget contains an AwesomeDialog widget that displays an error message if the user does not have access to view the details of a particular debitur.

🔨 refactor(home_search_nik.dart): extract search form and search result widgets
The search form and search result widgets are now extracted into their own files to improve code organization and readability. This also allows for easier maintenance and testing of these widgets. No functionality has been changed, only the structure of the code has been refactored.

🔥 refactor(home_search_nik.dart): remove unused NikSearchResult and NikSearchForm classes
The NikSearchResult and NikSearchForm classes are not used in the HomeSearchNik widget and can be safely removed to improve code readability and maintainability.
